backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
# psycopg2-binary  <- Aktifkan jika pakai PostgreSQL (production)
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
slowapi==0.1.9
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
